+++ packages/espressocash_app/lib/features/profile/screens/manage_profile_screen.dart
@@ -9,18 +9,17 @@ import '../../../../../ui/app_bar.dart';
 import '../../../../../ui/onboarding_screen.dart';
 import '../../../../../ui/text_field.dart';
 import '../../../../../ui/theme.dart';
+import '../../../core/email.dart';
 import '../../../core/file_manager.dart';
 import '../../../di.dart';
 import '../../../routes.gr.dart';
 import '../../../ui/back_button.dart';
-import '../../../ui/colors.dart';
 import '../../../ui/dialogs.dart';
 import '../../../ui/loader.dart';
+import '../../country_picker/models/country.dart';
+import '../../country_picker/widgets/country_picker.dart';
 import '../data/profile_repository.dart';
-import '../models/country.dart';
-import '../models/profile.dart';
 import '../widgets/pick_profile_picture.dart';
-import 'country_picker_screen.dart';
 
 @RoutePage()
 class ManageProfileScreen extends StatefulWidget {
@@ -39,6 +38,7 @@ class ManageProfileScreen extends StatefulWidget {
 
 class _ManageProfileScreenState extends State<ManageProfileScreen> {
   final _nameController = TextEditingController();
+  final _emailController = TextEditingController();
   Country? _country;
   File? _photo;
 
@@ -46,15 +46,14 @@ class _ManageProfileScreenState extends State<ManageProfileScreen> {
   void initState() {
     super.initState();
 
-    final profile = sl<ProfileRepository>().profile;
+    final repository = sl<ProfileRepository>();
 
-    _nameController
-      ..addListener(() => setState(() {}))
-      ..text = profile.firstName;
+    _nameController.text = repository.firstName;
+    _emailController.text = repository.email;
 
-    _photo = profile.photoPath?.let(File.new);
+    _photo = repository.photoPath?.let(File.new);
 
-    final country = profile.country;
+    final country = repository.country;
     if (country != null) {
       _country = Country.findByCode(country);
     }
@@ -63,30 +62,21 @@ class _ManageProfileScreenState extends State<ManageProfileScreen> {
   @override
   void dispose() {
     _nameController.dispose();
+    _emailController.dispose();
     super.dispose();
   }
-
-  void _handleCountryPressed() => context.router.push<Country>(
-        CountryPickerScreen.route(
-          initial: _country,
-          onSubmitted: (country) {
-            context.router.pop();
-            setState(() => _country = country);
-          },
-        ),
-      );
 
   void _handleSubmitted() => runWithLoader(context, () async {
         try {
           final photo = await _photo?.let(sl<FileManager>().copyToAppDir);
 
-          sl<ProfileRepository>().profile = Profile(
-            firstName: _nameController.text,
-            country: _country?.code,
-            photoPath: photo?.path,
-          );
-
           if (!mounted) return;
+
+          sl<ProfileRepository>()
+            ..firstName = _nameController.text
+            ..country = _country?.code
+            ..photoPath = photo?.path
+            ..email = _emailController.text;
 
           widget.onSubmitted();
         } on Exception catch (e) {
@@ -96,15 +86,21 @@ class _ManageProfileScreenState extends State<ManageProfileScreen> {
         }
       });
 
-  bool get _isValid => _nameController.text.isNotEmpty;
+  bool get _isValid =>
+      _nameController.text.isNotEmpty &&
+      _emailController.text.isValidEmail &&
+      _country != null;
 
   @override
   Widget build(BuildContext context) => CpTheme.dark(
         child: Scaffold(
           body: OnboardingScreen(
-            footer: OnboardingFooterButton(
-              text: context.l10n.save,
-              onPressed: _isValid ? _handleSubmitted : null,
+            footer: ListenableBuilder(
+              listenable: Listenable.merge([_nameController, _emailController]),
+              builder: (context, child) => OnboardingFooterButton(
+                text: context.l10n.save,
+                onPressed: _isValid ? _handleSubmitted : null,
+              ),
             ),
             children: [
               CpAppBar(
@@ -127,50 +123,33 @@ class _ManageProfileScreenState extends State<ManageProfileScreen> {
                   backgroundColor: Colors.white,
                 ),
               ),
-              _CountryPickerItem(
-                country: _country,
-                onTap: _handleCountryPressed,
+              OnboardingPadding(
+                child: CpTextField(
+                  margin: const EdgeInsets.only(top: 16),
+                  placeholder: context.l10n.yourEmailPlaceholder,
+                  controller: _emailController,
+                  backgroundColor: Colors.white,
+                ),
+              ),
+              const SizedBox(height: 8),
+              OnboardingPadding(
+                child: Padding(
+                  padding: const EdgeInsets.symmetric(horizontal: 24),
+                  child: Text(
+                    context.l10n.yourEmailDisclaimer,
+                    textAlign: TextAlign.center,
+                    style: const TextStyle(color: Colors.white, fontSize: 13),
+                  ),
+                ),
+              ),
+              const SizedBox(height: 40),
+              OnboardingPadding(
+                child: CountryPicker(
+                  country: _country,
+                  onSubmitted: (country) => setState(() => _country = country),
+                ),
               ),
             ],
-          ),
-        ),
-      );
-}
-
-class _CountryPickerItem extends StatelessWidget {
-  const _CountryPickerItem({
-    this.country,
-    required this.onTap,
-  });
-  final Country? country;
-  final VoidCallback onTap;
-
-  @override
-  Widget build(BuildContext context) => OnboardingPadding(
-        child: Container(
-          margin: const EdgeInsets.only(top: 16),
-          decoration: const ShapeDecoration(
-            color: CpColors.darkBackground,
-            shape: StadiumBorder(),
-          ),
-          child: ListTile(
-            contentPadding:
-                const EdgeInsets.symmetric(horizontal: 24, vertical: 10),
-            onTap: onTap,
-            title: Text(
-              country?.name ?? context.l10n.countryOfResidence,
-              style: const TextStyle(
-                fontWeight: FontWeight.normal,
-                fontSize: 20,
-                color: Colors.white,
-                height: 1.2,
-              ),
-            ),
-            trailing: const Icon(
-              Icons.keyboard_arrow_down_outlined,
-              color: Colors.white,
-              size: 34,
-            ),
           ),
         ),
       );