+++ packages/espressocash_app/lib/features/ramp/src/widgets/partners/ramp_network.dart
@@ -2,18 +2,20 @@ import 'package:flutter/material.dart';
 import 'package:ramp_flutter/configuration.dart';
 import 'package:ramp_flutter/ramp_flutter.dart';
 
-import '../../config.dart';
-import '../balances/widgets/context_ext.dart';
+import '../../../../../config.dart';
+import '../../../../balances/widgets/context_ext.dart';
+import '../../models/profile_data.dart';
 
 extension BuildContextExt on BuildContext {
   void launchRampNetworkOnRamp({
-    required String countryCode,
+    required ProfileData profile,
     required String address,
   }) {
     final configuration = _defaultConfiguration
-      ..selectedCountryCode = countryCode
+      ..selectedCountryCode = profile.country.code
       ..defaultFlow = 'ONRAMP'
-      ..userAddress = address;
+      ..userAddress = address
+      ..userEmailAddress = profile.email;
 
     RampFlutter()
       ..onRampClosed = notifyBalanceAffected